fixed the stockmanagement (#116)

+++ Models/BrandAmount.cs
@@ -11,7 +11,7 @@ namespace VaccineAPI.Models
         public int Count { get; set; }
 
         // public string SupName { get; set; } // Supplier Name
-        public int PurchasedAmt { get; set; } // Purchased Vaccine Amount
+        public decimal PurchasedAmt { get; set; } // Purchased Vaccine Amount
         // public bool IsPaid { get; set; } // Payment Status
 
         public long BrandId { get; set; }
